+++ evaluations/typeDef.js
@@ -25,6 +25,7 @@ const typeDefs = gql`
     competencies: [EvaluationCompetency!]
     course: Course!
     created: DateTime! @auth(requires: TEACHER)
+    data: JSONObject
     date: DateTime!
     evaluator: User
     explanation: String
@@ -34,6 +35,7 @@ const typeDefs = gql`
     learner: User!
     note: String @auth(requires: TEACHER)
     published: DateTime
+    requested: DateTime
     status: EvaluationStatus!
   }
 
@@ -42,6 +44,7 @@ const typeDefs = gql`
     evaluations(
       assessment: ID
       courseCode: ID
+      instance: ID
       learner: ID
       offset: Int
       limit: Int